Refactor ContactSection and HeroSection components; introduce Navbar component for improved navigation. Update contact methods and styles in ContactSection, and enhance layout in MentionsLegales page by integrating Navbar for consistent navigation across the application.
