+++ src/app/mentions-legales/page.tsx
@@ -1,43 +1,49 @@
+import Navbar from '../components/Navbar'
+
 export default function MentionsLegales() {
     return (
-        <main className="min-h-screen pt-32 pb-12 px-4">
-            <div className="mx-auto max-w-3xl">
-                <h1 className="text-4xl font-bold mb-8">Mentions légales</h1>
+        <>
+            <Navbar />
+            <main className="min-h-screen pt-32 pb-12 px-4">
+                <div className="mx-auto max-w-3xl">
+                    <h1 className="text-4xl font-bold mb-8">Mentions légales</h1>
 
-                <section className="mb-8">
-                    <h2 className="text-2xl font-semibold mb-4">1. Informations légales</h2>
-                    <p className="text-gray-600 mb-4">
-                        Prysm<br />
-                        [Adresse]<br />
-                        [Numéro SIRET]<br />
-                        Email : [email]<br />
-                        Téléphone : [téléphone]
-                    </p>
-                </section>
+                    <section className="mb-8">
+                        <h2 className="text-2xl font-semibold mb-4">1. Informations légales</h2>
+                        <p className="text-gray-600 mb-4">
+                            Prysm<br />
+                            20 rue Ramonet<br />
+                            33000 Bordeaux<br />
+                            Email : contact@prysm.studio
+                        </p>
+                    </section>
 
-                <section className="mb-8">
-                    <h2 className="text-2xl font-semibold mb-4">2. Hébergement</h2>
-                    <p className="text-gray-600">
-                        Ce site est hébergé par :<br />
-                        [Nom de l&apos;hébergeur]<br />
-                        [Adresse de l&apos;hébergeur]
-                    </p>
-                </section>
+                    <section className="mb-8">
+                        <h2 className="text-2xl font-semibold mb-4">2. Hébergement</h2>
+                        <p className="text-gray-600">
+                            Ce site est hébergé par :<br />
+                            Vercel Inc.<br />
+                            340 S Lemon Ave #4133<br />
+                            Walnut, CA 91789<br />
+                            États-Unis
+                        </p>
+                    </section>
 
-                <section className="mb-8">
-                    <h2 className="text-2xl font-semibold mb-4">3. Propriété intellectuelle</h2>
-                    <p className="text-gray-600">
-                        L&apos;ensemble de ce site relève de la législation française et internationale sur le droit d&apos;auteur et la propriété intellectuelle. Tous les droits de reproduction sont réservés, y compris pour les documents téléchargeables et les représentations iconographiques et photographiques.
-                    </p>
-                </section>
+                    <section className="mb-8">
+                        <h2 className="text-2xl font-semibold mb-4">3. Propriété intellectuelle</h2>
+                        <p className="text-gray-600">
+                            L&apos;ensemble de ce site relève de la législation française et internationale sur le droit d&apos;auteur et la propriété intellectuelle. Tous les droits de reproduction sont réservés, y compris pour les documents téléchargeables et les représentations iconographiques et photographiques.
+                        </p>
+                    </section>
 
-                <section className="mb-8">
-                    <h2 className="text-2xl font-semibold mb-4">4. Protection des données personnelles</h2>
-                    <p className="text-gray-600">
-                        Conformément à la loi &quot;Informatique et Libertés&quot; du 6 janvier 1978 modifiée, et au Règlement Général sur la Protection des Données (RGPD), vous disposez d&apos;un droit d&apos;accès, de rectification et de suppression des données vous concernant.
-                    </p>
-                </section>
-            </div>
-        </main>
+                    <section className="mb-8">
+                        <h2 className="text-2xl font-semibold mb-4">4. Protection des données personnelles</h2>
+                        <p className="text-gray-600">
+                            Conformément à la loi &quot;Informatique et Libertés&quot; du 6 janvier 1978 modifiée, et au Règlement Général sur la Protection des Données (RGPD), vous disposez d&apos;un droit d&apos;accès, de rectification et de suppression des données vous concernant.
+                        </p>
+                    </section>
+                </div>
+            </main>
+        </>
     )
 }
